+++ CHANGELOG.md
@@ -18,6 +18,8 @@
 - Update `config.transport.proxy` to allow String and URI values as previously supported by `sentry-ruby` versions <= 4.8 using Faraday
   - Fixes [#1782](https://github.com/getsentry/sentry-ruby/issues/1782)
 - Register SentryContextClientMiddleware on sidekiq workers [#1774](https://github.com/getsentry/sentry-ruby/pull/1774)
+- Add request env to sampling context when using `sentry-rails` [#1792](https://github.com/getsentry/sentry-ruby/pull/1792)
+  - Fixes [#1791](https://github.com/getsentry/sentry-ruby/issues/1791)
 
 ### Refactoring