+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## Unreleased
+## 5.3.0
 
 ### Features